refactor(voice): isolate provider implementations
